dnsspoofer/requirements.txt
scapy
colorama
netfilterqueue
fire
python-whois
